Feat: Splash Animation
