+++ app/build.gradle
@@ -78,4 +78,8 @@ dependencies {
 
     // Infinity Scroll
     implementation 'androidx.swiperefreshlayout:swiperefreshlayout:1.1.0'
+
+    // Splash Animation
+    // implementation 'com.airbnb.andorid:lottie:2.1.0'
+    implementation 'com.airbnb.android:lottie:1.5.3'
 }
