+++ app/src/main/java/com/example/geeksasaeng/SplashActivity.kt
@@ -18,6 +18,10 @@ class SplashActivity: BaseActivity<ActivitySplashBinding>(ActivitySplashBinding:
     var password: String? = null
 
     override fun initAfterBinding() {
+        // Splash Animation 적용
+        binding.splashAnimView.imageAssetsFolder = "images"
+        binding.splashAnimView.playAnimation()
+
         val handler = Handler(Looper.getMainLooper())
 
         jwt = getJwt()
@@ -39,7 +43,6 @@ class SplashActivity: BaseActivity<ActivitySplashBinding>(ActivitySplashBinding:
             }
         }, 1500)
     }
-
 
     private fun login() {
         val loginDataService = LoginDataService()
