+++ app/src/main/res/layout/activity_splash.xml
@@ -5,52 +5,20 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_gravity="center_vertical"
+    android:background="@color/white"
     tools:context=".MainActivity">
 
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="39dp"
+    <com.airbnb.lottie.LottieAnimationView
+        android:id="@+id/splash_anim_view"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        app:lottie_autoPlay="true"
+        app:lottie_fileName="data.json"
+        app:lottie_loop="true"
+        app:lottie_imageAssetsFolder="images"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent">
-
-        <ImageView
-            android:id="@+id/splash_logo"
-            android:layout_width="209dp"
-            android:layout_height="209dp"
-            android:src="@drawable/ic_logo"
-            android:scaleType="fitXY"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent" />
-
-        <TextView
-            android:id="@+id/splash_sub_tv"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="쉽고 편리한 연결 메이트"
-            android:textColor="@color/black"
-            android:textSize="18sp"
-            android:fontFamily="@font/gmarket_medium"
-            app:layout_constraintTop_toBottomOf="@id/splash_logo"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent" />
-
-        <TextView
-            android:id="@+id/splash_main_tv"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="긱사생"
-            android:textColor="@color/black"
-            android:textSize="32sp"
-            android:fontFamily="@font/gmarket_medium"
-            android:layout_marginTop="19dp"
-            app:layout_constraintTop_toBottomOf="@id/splash_sub_tv"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
+        app:layout_constraintBottom_toBottomOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
